Collect transforms in new module T.P.Transforms. [API change]
Bundles all document transformations, i.e., functions that run
predefined filters on Pandoc documents.
