+++ pandoc.cabal
@@ -651,6 +651,7 @@ library
                    Text.Pandoc.Class.IO,
                    Text.Pandoc.Citeproc,
                    Text.Pandoc.Chunks,
+                   Text.Pandoc.Transforms,
                    Text.Pandoc.Version
   other-modules:   Text.Pandoc.App.CommandLineOptions,
                    Text.Pandoc.App.Input,
